fix test for app component
