+++ src/main/js/src/App.test.js
@@ -1,9 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import { StaticRouter } from 'react-router-dom';
 
-xit('renders without crashing', () => {
+it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+
+  const app = (
+    <StaticRouter>
+      <App />
+    </StaticRouter>
+  );
+
+  ReactDOM.render(app, div);
   ReactDOM.unmountComponentAtNode(div);
 });
