+++ src/main/js/src/hoc/withErrorHandler/withErrorHandler.js
@@ -7,7 +7,7 @@ const withErrorHandler = (WrappedComponent, axios) => {
       error: false,
     };
 
-    componentWillMount() {
+    componentDidMount() {
       this.reqInterceptor = axios.interceptors.request.use(req => {
         this.setState({ error: null });
         return req;
