+++ src/main/js/src/index.js
@@ -1,38 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
-import { Provider } from 'react-redux';
-import thunk from 'redux-thunk';
 
+import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import organizerReducer from './store/reducers/organizer';
-import examSessionReducer from './store/reducers/examSession';
-import registrationReducer from './store/reducers/registration';
-import './index.css';
 
-const rootReducer = combineReducers({
-  org: organizerReducer,
-  exam: examSessionReducer,
-  reg: registrationReducer,
-});
-
-// To enable https://github.com/zalmoxisus/redux-devtools-extension
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-const store = createStore(
-  rootReducer,
-  composeEnhancers(applyMiddleware(thunk)),
-);
-
-const app = (
-  <Provider store={store}>
-    <BrowserRouter basename="/yki">
-      <App />
-    </BrowserRouter>
-  </Provider>
-);
-
-ReactDOM.render(app, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
 serviceWorker.unregister();
